Move New swap code from page to NewSwapWidget
