+++ src/cow-react/modules/swap/pure/SwapForm/index.cosmos.tsx
@@ -1,5 +1,5 @@
 import { SwapForm } from 'cow-react/modules/swap/pure/SwapForm/index'
-import { SwapFormProps } from 'cow-react/pages/NewSwap/typings'
+import { SwapFormProps } from 'cow-react/modules/swap/containers/NewSwapWidget/typings'
 import { defaultCurrencyInputPanelProps } from 'cow-react/common/pure/CurrencyInputPanel/defaultCurrencyInputProps'
 import { GNO } from 'constants/tokens'
 import { SupportedChainId } from 'constants/chains'
